Refactor the oracle connector to use base SqlConnector
