+++ lib/oracle.js
@@ -12,6 +12,7 @@ try {
   }
 }
 var SqlConnector = require('loopback-connector').SqlConnector;
+var ParameterizedSQL = SqlConnector.ParameterizedSQL;
 var util = require('util');
 var async = require('async');
 var debug = require('debug')('loopback:connector:oracle');
@@ -134,11 +135,8 @@ Oracle.prototype.connect = function (callback) {
  * @param {String[]} params The parameter values for the SQL statement.
  * @param {Function} [callback] The callback after the SQL statement is executed.
  */
-var open = 0;
-Oracle.prototype.executeSQL = function (sql, params, callback) {
+Oracle.prototype.executeSQL = function (sql, params, options, callback) {
   var self = this;
-  var time = Date.now();
-  var log = self.log;
 
   if (self.settings.debug) {
     if (params && params.length > 0) {
@@ -180,146 +178,53 @@ Oracle.prototype.executeSQL = function (sql, params, callback) {
     });
   });
   */
-  self.pool.execute(sql, params, callback);
-};
-
-/**
- *
- * Execute the SQL statement synchronously.
- *
- * @param {String} sql The SQL statement
- * @param {String[]} params An array of parameters
- *
- */
-Oracle.prototype.executeSQLSync = function (sql, params) {
-  var self = this;
-  if (self.settings.debug) {
-    if (params && params.length > 0) {
-      self.debug('SQL: %s \nParameters: %j', sql, params);
-    } else {
-      self.debug('SQL: %s', sql);
+  self.pool.execute(sql, params, function(err, result) {
+    if (err) {
+      self.debug('Error: %j', err);
     }
-  }
-
-  if (!self.pool) {
-    self.pool = self.driver.createConnectionPoolSync(self.settings);
-    if (self.settings.debug) {
-      debugConnection('Connection acquired: ', self.pool.getInfo());
+    if (result) {
+      self.debug('Result: %j', result);
     }
-  }
-  var connection = self.pool.getConnectionSync();
-  try {
-    var data = connection.executeSync(sql, params);
-    if (self.settings.debug && data) {
-      self.debug("%j", data);
-    }
-    connection.close();
-    if (self.settings.debug) {
-      debugConnection('Connection released: ', self.pool.getInfo());
-    }
-    connection = null;
-    return data;
-  } catch (err) {
-    if (connection) {
-      connection.close();
-      if (self.settings.debug) {
-        self.debug('Connection released: ', self.pool.getInfo());
-      }
-    }
-    throw err;
-  }
-
-};
-
-/*
- * Check if the connection is in progress.
- * @private
- */
-function stillConnecting(dataSource, obj, args) {
-  if (dataSource.connected) {
-    // already connected
-    return false;
-  }
-
-  var method = args.callee;
-  // Set up a callback after the connection is established to continue the method call
-  dataSource.once('connected', function () {
-    method.apply(obj, [].slice.call(args));
+    callback(err, result);
   });
-  if (!dataSource.connecting) {
-    dataSource.connect();
-  }
-  return true;
-}
-
-/**
- * Execute a SQL statement with the given parameters
- *
- * @param {String} sql The SQL statement
- * @param {String[]} params An array of parameter values
- * @param {Function} [callback] The callback function
- */
-Oracle.prototype.query = function (sql, params, callback) {
-  if (stillConnecting(this.dataSource, this, arguments)) return;
-
-  if (!callback && typeof params === 'function') {
-    callback = params;
-    params = [];
-  }
-
-  params = params || [];
-
-  var cb = callback || function (err, result) {
-  };
-  this.executeSQL(sql, params, cb);
 };
 
 /**
- * Run a query synchronously
- *
- * @param {String} sql The SQL statement
- * @param {String[]} params An array of parameter values
- * @returns {*} The results
+ * Get the place holder in SQL for values, such as :1 or ?
+ * @param {String} key Optional key, such as 1 or id
+ * @returns {String} The place holder
  */
-Oracle.prototype.querySync = function (sql, params) {
-  params = params || [];
-
-  return this.executeSQLSync(sql, params);
+Oracle.prototype.getPlaceholderForValue = function(key) {
+  return ':' + key;
 };
 
-/**
- * Count the number of instances for the given model
- *
- * @param {String} model The model name
- * @param {Function} [callback] The callback function
- * @param {Object} filter The filter for where
- *
- */
-Oracle.prototype.count = function count(model, callback, filter) {
-  this.query('SELECT count(*) as "cnt"  FROM '
-    + ' ' + this.toFilter(model, filter && {where: filter}), function (err, data) {
-    if (err) return callback(err);
-    callback(err, data && data[0] && data[0].cnt);
-  }.bind(this));
+Oracle.prototype.getCountForAffectedRows = function(model, info) {
+  return info && info.updateCount;
 };
 
-/**
- * Delete instances for the given model
- *
- * @param {String} model The model name
- * @param {Object} [where] The filter for where
- * @param {Function} [callback] The callback function
- *
- */
-Oracle.prototype.destroyAll = function destroyAll(model, where, callback) {
-  if (!callback && 'function' === typeof where) {
-    callback = where;
-    where = undefined;
+Oracle.prototype.getInsertedId = function(model, info) {
+  return info && info.returnParam;
+};
+
+Oracle.prototype.buildInsertDefaultValues = function(model, data, options) {
+  // Oracle doesn't like empty column/value list
+  var idCol = this.idColumnEscaped(model);
+  return '(' + idCol + ') VALUES(DEFAULT)';
+};
+
+Oracle.prototype.buildInsertReturning = function(model, data, options) {
+  var modelDef = this.getModelDefinition(model);
+  var type = modelDef.properties[this.idName(model)].type;
+  var outParam = null;
+  if (type === String) {
+    outParam = new this.driver.OutParam(this.driver.OCCISTRING);
+  } else {
+    outParam = new this.driver.OutParam();
   }
-  this.query('DELETE  FROM '
-    + ' ' + this.toFilter(model, where && {where: where}), function (err, info) {
-    callback && callback(err, {count: info && info.updateCount});
-  }.bind(this));
+  var params = [outParam];
+  var returningStmt = new ParameterizedSQL('RETURNING ' +
+    this.idColumnEscaped(model) + ' into ?', params);
+  return returningStmt;
 };
 
 /**
@@ -329,142 +234,21 @@ Oracle.prototype.destroyAll = function destroyAll(model, where, callback) {
  * @param {Object} data The model instance data
  * @param {Function} [callback] The callback function
  */
-Oracle.prototype.create = function (model, data, callback) {
-  var fields = this.toFields(model, data, true);
-  var sql = 'INSERT INTO ' + this.tableEscaped(model) + '';
-  if (fields) {
-    sql += ' ' + fields.sql;
-  } else {
-    sql += ' VALUES ()';
-  }
-  sql += ' RETURNING ' + this.idColumnEscaped(model) + ' into :' + (fields.params.length + 1);
-
-  var type = this._models[model].properties[this.idName(model)].type;
-  var outParam = null;
-  if (type === String) {
-    outParam = new this.driver.OutParam(this.driver.OCCISTRING);
-  } else {
-    outParam = new this.driver.OutParam();
-  }
-
-  var params = fields.params;
-  params.push(outParam);
-  this.query(sql, params, function (err, info) {
+Oracle.prototype.create = function(model, data, options, callback) {
+  var self = this;
+  var stmt = this.buildInsert(model, data, options);
+  this.execute(stmt.sql, stmt.params, function(err, info) {
     if (err) {
       if (err.toString().indexOf('ORA-00001: unique constraint') >= 0) {
         // Transform the error so that duplicate can be checked using regex
         err = new Error(err.toString() + '. Duplicate id detected.');
       }
-      return callback(err);
-    }
-    callback(err, info && info.returnParam);
-  });
-};
-
-/**
- * Save the model instance to Oracle DB
- * @param {String} model The model name
- * @param {Object} data The model instance data
- * @param {Function} [callback] The callback function
- */
-Oracle.prototype.save = function (model, data, callback) {
-  var self = this;
-  var fields = this.toFields(model, data);
-  var sql = 'UPDATE ' + this.tableEscaped(model) + ' SET ' + fields.sql;
-
-  sql += ' WHERE ';
-  var idNames = this.idNames(model);
-  idNames.forEach(function (id, i) {
-    if (data[id] !== null && data[id] !== undefined) {
-      sql += ((i > 0) ? ' AND ' : '') + self.idColumnEscaped(model) + ' = ' + data[id];
+      callback(err);
     } else {
-      sql += ((i > 0) ? ' AND ' : '') + self.idColumnEscaped(model) + ' IS NULL';
+      var insertedId = self.getInsertedId(model, info);
+      callback(err, insertedId);
     }
   });
-
-  this.query(sql, fields.params, function (err) {
-    callback(err);
-  });
-};
-
-Oracle.prototype.update = Oracle.prototype.updateAll = function (model, where, data, callback) {
-  var fields = this.toFields(model, data);
-  var sql = 'UPDATE ' + this.tableEscaped(model) + ' SET ' + fields.sql;
-
-  sql += this.buildWhere(model, where);
-
-  this.query(sql, fields.params, function (err, info) {
-    callback && callback(err, {count: info && info.updateCount});
-  });
-};
-
-/*!
- * Build a list of column name/value pairs
- *
- * @param {String} The model name
- * @param {Object} The model instance data
- * @param {Boolean} forCreate Indicate if it's for creation
- */
-Oracle.prototype.toFields = function (model, data, forCreate) {
-  var result = {
-  };
-  var columns = [];
-  var fields = [];
-  var params = [];
-  var props = this._models[model].properties;
-  var self = this;
-
-  var index = 0;
-  if (forCreate) {
-    Object.keys(data).forEach(function (key) {
-      if (props[key]) {
-        if (!self.id(model, key)) {
-          if (props[key].type === String || props[key].type === Buffer) {
-            columns.push(self.columnEscaped(model, key));
-            fields.push(':' + (++index));
-            params.push(self.toDatabase(props[key], data[key], true));
-          } else {
-            columns.push(self.columnEscaped(model, key));
-            fields.push(self.toDatabase(props[key], data[key]));
-          }
-        } else {
-          if (data[key]) {
-            if (props[key].type === String || props[key].type === Buffer) {
-              columns.push(self.columnEscaped(model, key));
-              fields.push(':' + (++index));
-              params.push(self.toDatabase(props[key], data[key], true));
-            } else {
-              columns.push(self.columnEscaped(model, key));
-              fields.push(self.toDatabase(props[key], data[key]));
-            }
-          }
-        }
-      }
-    });
-    if (columns.length > 0) {
-      result.sql = '(' + columns.join(',') + ') VALUES (' + fields.join(',') + ')';
-    } else {
-      // Oracle doesn't like empty column/value list
-      var idName = self.columnEscaped(model, self.idName(model));
-      result.sql = '(' + idName + ') VALUES (default)';
-    }
-    result.params = params;
-    return result;
-  } else {
-    Object.keys(data).forEach(function (key) {
-      if (props[key]) {
-        if (props[key].type === String || props[key].type === Buffer) {
-          columns.push(self.columnEscaped(model, key) + '= ' + (':' + (++index)));
-          params.push(self.toDatabase(props[key], data[key], true));
-        } else {
-          columns.push(self.columnEscaped(model, key) + '= ' + self.toDatabase(props[key], data[key]));
-        }
-      }
-    });
-    result.sql = columns.join(',');
-    result.params = params;
-    return result;
-  }
 };
 
 function dateToOracle(val, dateOnly) {
@@ -497,103 +281,67 @@ function dateToOracle(val, dateOnly) {
   }
 
   if (dateOnly) {
-    return "to_date('" + dateStr + "', 'yyyy-mm-dd hh24:mi:ss')";
+    return new ParameterizedSQL(
+      "to_date(?,'yyyy-mm-dd hh24:mi:ss')", [dateStr]);
   } else {
-    return "to_timestamp('" + dateStr + "', 'yyyy-mm-dd hh24:mi:ss.ff3')";
+    return new ParameterizedSQL(
+      "to_timestamp(?,'yyyy-mm-dd hh24:mi:ss.ff3')", [dateStr]);
   }
 
 }
 
-/*!
- * Convert name/value to database value
- *
- * @param {String} prop The property name
- * @param {*} val The property value
- * @param {Boolean} raw
- */
-Oracle.prototype.toDatabase = function (prop, val, raw) {
-  if (val === null || val === undefined) {
-    // Oracle complains with NULLs in not null columns
+Oracle.prototype.toColumnValue = function(prop, val) {
+  if (val == null) {
+    // PostgreSQL complains with NULLs in not null columns
     // If we have an autoincrement value, return DEFAULT instead
-    if (prop.autoIncrement) {
-      return 'DEFAULT';
+    if (prop.autoIncrement || prop.id) {
+      return new ParameterizedSQL('DEFAULT');
     }
     else {
-      return raw ? null : 'NULL';
+      return null;
     }
   }
-  if (prop.type.name === 'Number') {
-    if (!val && val !== 0) {
-      if (prop.autoIncrement) {
-        return 'DEFAULT';
-      }
-      else {
-        return 'NULL';
-      }
+  if (prop.type === String) {
+    return String(val);
+  }
+  if (prop.type === Number) {
+    if (isNaN(val)) {
+      // Map NaN to NULL
+      return val;
     }
-    return escape(val);
+    return val;
   }
 
-  if (prop.type.name === 'Date' || prop.type.name === 'Timestamp') {
-    if (!val) {
-      if (prop.autoIncrement) {
-        return 'DEFAULT';
-      }
-      else {
-        return 'NULL';
-      }
-    }
-    if (!val.toUTCString) {
-      val = new Date(val);
-    }
-    return dateToOracle(val, prop.type.name === 'Date');
+  if (prop.type === Date || prop.type.name === 'Timestamp') {
+    return dateToOracle(val, prop.type === Date);
   }
 
   // Oracle support char(1) Y/N
-  if (prop.type.name === 'Boolean') {
+  if (prop.type === Boolean) {
     if (val) {
-      return "'Y'";
+      return 'Y';
     } else {
-      return "'N'";
+      return 'N';
     }
   }
 
-  if (typeof val === 'object') {
-    val = JSON.stringify(val);
-  }
-
-  return raw ? val : escape(val.toString());
-
+  return this.serializeObject(val);
 };
 
-/*!
- * Convert the data from database to JSON
- *
- * @param {String} model The model name
- * @param {Object} data The data from DB
- */
-Oracle.prototype.fromDatabase = function (model, data) {
-  if (!data) return null;
-  var props = this._models[model].properties;
-  var json = {};
-  for (var p in props) {
-    var key = this.column(model, p);
-    var val = data[key];
-    if (val === undefined) {
-      continue;
-    }
-    var prop = props[p];
-    if (prop && prop.type && prop.type.name === 'Boolean') {
-      json[p] = (val === 'Y' || val === 'y' || val === 'T' || val === 't'
-        || val === '1');
+Oracle.prototype.fromColumnValue = function(prop, val) {
+  if (val == null) {
+    return val;
+  }
+  var type = prop && prop.type;
+  if (type === Boolean) {
+    if (typeof val === 'boolean') {
+      return val;
     } else {
-      json[p] = val;
+      return (val === 'Y' || val === 'y' || val === 'T' ||
+      val === 't' || val === '1');
     }
   }
-  if (this.settings.debug) {
-    this.debug('JSON data: %j', json);
-  }
-  return json;
+  return val;
 };
 
 /*!
@@ -617,1304 +365,120 @@ Oracle.prototype.escapeName = function (name) {
   if (!name) {
     return name;
   }
-  return '"' + name.replace(/\./g, '"."').toUpperCase() + '"';
+  return '"' + name.replace(/\./g, '"."') + '"';
 };
 
-Oracle.prototype.schemaName = function (model) {
-  // Check if there is a 'schema' property for mssql
-  var dbMeta = this._models[model].settings && this._models[model].settings.oracle;
-  var schemaName = (dbMeta && (dbMeta.schema || dbMeta.schemaName))
-    || this.settings.schema || this.settings.user;
-  return schemaName;
-};
 
 Oracle.prototype.tableEscaped = function (model) {
-  var schemaName = this.schemaName(model);
+  var schemaName = this.schema(model);
   if (schemaName && schemaName !== this.settings.user) {
-    return this.escapeName(schemaName) + '.' + this.escapeName(this.table(model));
+    return this.escapeName(schemaName) + '.' +
+      this.escapeName(this.table(model));
   } else {
     return this.escapeName(this.table(model));
   }
 };
-/*!
- * Get a list of columns based on the fields pattern
- *
- * @param {String} model The model name
- * @param {Object|String[]} props Fields pattern
- * @returns {String}
- */
-Oracle.prototype.getColumns = function (model, props) {
-  var cols = this._models[model].properties;
-  var self = this;
-  var keys = Object.keys(cols);
-  if (Array.isArray(props) && props.length > 0) {
-    // No empty array, including all the fields
-    keys = props;
-  } else if ('object' === typeof props && Object.keys(props).length > 0) {
-    // { field1: boolean, field2: boolean ... }
-    var included = [];
-    var excluded = [];
-    keys.forEach(function (k) {
-      if (props[k]) {
-        included.push(k);
-      } else if ((k in props) && !props[k]) {
-        excluded.push(k);
+
+Oracle.prototype.buildExpression =
+  function(columnName, operator, columnValue, propertyValue) {
+    if (propertyValue instanceof RegExp) {
+      columnValue = "'" + propertyValue.source + "'";
+      if (propertyValue.ignoreCase) {
+        return new ParameterizedSQL(columnName + ' ~* ?', [columnValue]);
+      } else {
+        return new ParameterizedSQL(columnName + ' ~ ?', [columnValue]);
       }
-    });
-    if (included.length > 0) {
-      keys = included;
-    } else if (excluded.length > 0) {
-      excluded.forEach(function (e) {
-        var index = keys.indexOf(e);
-        keys.splice(index, 1);
-      });
     }
+    switch(operator) {
+      case 'like':
+        return new ParameterizedSQL({
+          sql: columnName + " LIKE ? ESCAPE '\\'",
+          params: [columnValue]
+        });
+      case 'nlike':
+        return new ParameterizedSQL({
+          sql: columnName + " NOT LIKE ? ESCAPE '\\'",
+          params: [columnValue]
+        });
+      default:
+        // Invoke the base implementation of `buildExpression`
+        var exp = this.invokeSuper('buildExpression',
+          columnName, operator, columnValue, propertyValue);
+        return exp;
+    }
+  };
+
+function buildLimit(limit, offset) {
+  if (isNaN(offset)) {
+    offset = 0;
   }
-  var names = keys.map(function (c) {
-    return self.columnEscaped(model, c);
-  });
-  return names.join(', ');
+  var sql = 'OFFSET ' + offset + ' ROWS';
+  if (limit >= 0) {
+    sql += ' FETCH NEXT ' + limit + ' ROWS ONLY';
+  }
+  return sql;
+}
+
+Oracle.prototype.applyPagination =
+  function(model, stmt, filter) {
+    var offset = filter.offset || filter.skip || 0;
+    if (this.settings.supportsOffsetFetch) {
+      // Oracle 12.c or later
+      var limitClause = buildLimit(filter.limit, filter.offset || filter.skip);
+      return stmt.merge(limitClause);
+    } else {
+      var paginatedSQL = 'SELECT * FROM (' + stmt.sql + ' ' +
+        ')' + ' ' + ' WHERE R > ' + offset;
+
+      if (filter.limit !== -1) {
+        paginatedSQL += ' AND R <= ' + (offset + filter.limit);
+      }
+
+      stmt.sql = paginatedSQL + ' ';
+      return stmt;
+    }
+  };
+
+Oracle.prototype.buildColumnNames = function(model, filter) {
+  var columnNames = this.invokeSuper('buildColumnNames', model, filter);
+  if (filter.limit || filter.offset || filter.skip) {
+    var orderBy = this.buildOrderBy(model, filter.order);
+    columnNames += ',ROW_NUMBER() OVER' + ' (' + orderBy + ') R';
+  }
+  return columnNames;
 };
 
-/**
- * Find matching model instances by the filter
- *
- * @param {String} model The model name
- * @param {Object} filter The filter
- * @param {Function} [callback] The callback function
- */
-Oracle.prototype.all = function all(model, filter, callback) {
-  // Order by id if no order is specified
-  filter = filter || {};
+Oracle.prototype.buildSelect = function(model, filter, options) {
   if (!filter.order) {
     var idNames = this.idNames(model);
     if (idNames && idNames.length) {
       filter.order = idNames;
     }
   }
-  this.query('SELECT ' + this.getColumns(model, filter.fields) + '  FROM '
-    + this.toFilter(model, filter), function (err, data) {
-    if (err) {
-      return callback(err, []);
+
+  var selectStmt = new ParameterizedSQL('SELECT ' +
+    this.buildColumnNames(model, filter) +
+    ' FROM ' + this.tableEscaped(model)
+  );
+
+  if (filter) {
+
+    if (filter.where) {
+      var whereStmt = this.buildWhere(model, filter.where);
+      selectStmt.merge(whereStmt);
     }
-    if (data) {
-      for (var i = 0; i < data.length; i++) {
-        data[i] = this.fromDatabase(model, data[i]);
-      }
-    }
-    if (filter && filter.include) {
-      this._models[model].model.include(data, filter.include, callback);
+
+    if (filter.limit || filter.skip || filter.offset) {
+      selectStmt = this.applyPagination(
+        model, selectStmt, filter);
     } else {
-      callback(null, data);
-    }
-  }.bind(this));
-};
-
-function getPagination(filter) {
-  var pagination = [];
-  if (filter && (filter.limit || filter.offset || filter.skip)) {
-    var offset = Number(filter.offset);
-    if (!offset) {
-      offset = Number(filter.skip);
-    }
-    if (offset) {
-      pagination.push('R >= ' + (offset + 1));
-    } else {
-      offset = 0;
-    }
-    var limit = Number(filter.limit);
-    if (limit) {
-      pagination.push('R <= ' + (offset + limit));
-    }
-  }
-  return pagination;
-}
-
-Oracle.prototype.buildWhere = function (model, conds) {
-  var where = this._buildWhere(model, conds);
-  if (where) {
-    return ' WHERE ' + where;
-  } else {
-    return '';
-  }
-};
-
-Oracle.prototype._buildWhere = function (model, where) {
-  if (!where) {
-    return '';
-  }
-  var self = this;
-  var props = self._models[model].properties;
-  var fields = [];
-  var conds = where;
-  if (typeof conds === 'string') {
-    fields.push(conds);
-  } else if (util.isArray(conds)) {
-    var query = conds.shift().replace(/\?/g, function (s) {
-      return escape(conds.shift());
-    });
-    fields.push(query);
-  } else {
-    Object.keys(conds).forEach(function (key) {
-      if (key === 'and' || key === 'or') {
-        var clauses = conds[key];
-        if (Array.isArray(clauses)) {
-          clauses = clauses.map(function (c) {
-            return '(' + self._buildWhere(model, c) + ')';
-          });
-          return fields.push(clauses.join(' ' + key.toUpperCase() + ' '));
-        }
-        // The value is not an array, fall back to regular fields
-      }
-      var sqlCond = null;
-      if (where[key] && where[key].constructor.name === 'RegExp') {
-        var regex = where[key];
-        sqlCond = self.columnEscaped(model, key);
-
-        if (regex.ignoreCase) {
-          sqlCond += ' ~* ';
-        } else {
-          sqlCond += ' ~ ';
-        }
-        sqlCond += "'" + regex.source + "'";
-        fields.push(sqlCond);
-        return;
-      }
-      if (props[key]) {
-        var filterValue;
-        if (conds[key].constructor.name === 'Object') {
-          var operator = Object.keys(conds[key])[0];
-          var val = conds[key][operator];
-          if (operator === 'between') {
-            filterValue = self.toDatabase(props[key], val[0]) + ' AND ' +
-              self.toDatabase(props[key], val[1]);
-          } else if (operator === 'inq' || operator === 'nin') {
-            for (var i = 0; i < val.length; i++) {
-              val[i] = self.toDatabase(props[key], val[i]);
-            }
-            filterValue = val.join(',');
-          } else {
-            filterValue = self.toDatabase(props[key], val);
-          }
-          sqlCond = self.columnEscaped(model, key);
-          if ((operator === 'inq' || operator === 'nin') && filterValue.length === 0) {
-            fields.push(operator === 'inq' ? '1 = 2' : '1 = 1');
-            return true;
-          }
-          switch (operator) {
-            case 'gt':
-              sqlCond += ' > ';
-              break;
-            case 'gte':
-              sqlCond += ' >= ';
-              break;
-            case 'lt':
-              sqlCond += ' < ';
-              break;
-            case 'lte':
-              sqlCond += ' <= ';
-              break;
-            case 'between':
-              sqlCond += ' BETWEEN ';
-              break;
-            case 'inq':
-              sqlCond += ' IN ';
-              break;
-            case 'nin':
-              sqlCond += ' NOT IN ';
-              break;
-            case 'neq':
-              sqlCond += ' != ';
-              break;
-            case 'like':
-              sqlCond += ' LIKE ';
-              filterValue += " ESCAPE '\\'";
-              break;
-            case 'nlike':
-              sqlCond += ' NOT LIKE ';
-              filterValue += " ESCAPE '\\'";
-              break;
-            default:
-              sqlCond += ' ' + operator + ' ';
-              break;
-          }
-          sqlCond += (operator === 'inq' || operator === 'nin') ?
-            '(' + filterValue + ')' : filterValue;
-          fields.push(sqlCond);
-        } else {
-          filterValue = self.toDatabase(props[key], conds[key]);
-          if (filterValue === 'NULL') {
-            fields.push(self.columnEscaped(model, key) + ' IS ' + filterValue);
-          } else {
-            fields.push(self.columnEscaped(model, key) + ' = ' + filterValue);
-          }
-        }
-      }
-    });
-  }
-  return fields.join(' AND ');
-};
-
-/*!
- * Build the SQL clause
- * @param {String} model The model name
- * @param {Object} filter The filter
- * @returns {*}
- */
-Oracle.prototype.toFilter = function (model, filter) {
-  var self = this;
-  if (filter && typeof filter.where === 'function') {
-    return this.tableEscaped(model) + ' ' + filter.where();
-  }
-  if (!filter) {
-    return this.tableEscaped(model);
-  }
-  var out = '';
-
-  var where = self.buildWhere(model, filter.where);
-
-  if (where) {
-    out += where;
-  }
-
-  // First check the pagination requirements
-  // http://docs.oracle.com/cd/B19306_01/server.102/b14200/functions137.htm#i86310
-  var pagination = getPagination(filter);
-
-  if (filter.order) {
-    var order = filter.order;
-    if (typeof order === 'string') {
-      order = [order];
-    }
-    var orderBy = '';
-    filter.order = [];
-    for (var i = 0, n = order.length; i < n; i++) {
-      var t = order[i].split(/[\s]+/);
-      var field = t[0], dir = t[1];
-      filter.order.push(self.columnEscaped(model, field) + (dir ? ' ' + dir : ''));
-    }
-    orderBy = ' ORDER BY ' + filter.order.join(',');
-    if (pagination.length) {
-      // We need to apply ROWNUM around the ORDER BY
-      out = '(SELECT ROW_NUMBER() OVER (' + orderBy + ') R, '
-        + this.getColumns(model) + ' FROM ' + this.tableEscaped(model) + out + ') WHERE '
-        + pagination.join(' AND ');
-    } else {
-      out = this.tableEscaped(model) + ' ' + out + ' ' + orderBy;
-    }
-  } else {
-    if (pagination.length) {
-      // We need to apply ROWNUM around the ORDER BY
-      out = '(SELECT ROW_NUMBER() OVER (ORDER BY 1) R, ' + this.getColumns(model) + ' FROM '
-        + this.tableEscaped(model) + out + ') WHERE ' + pagination.join(' AND ');
-    } else {
-      out = this.tableEscaped(model) + ' ' + out;
-    }
-  }
-
-  return out;
-};
-
-/**
- * Check if a model instance exists by id
- * @param {String} model The model name
- * @param {*} id The id value
- * @param {Function} [callback] The callback function
- *
- */
-Oracle.prototype.exists = function (model, id, callback) {
-  var sql = 'SELECT 1 FROM ' +
-    this.tableEscaped(model);
-
-  if (id) {
-    sql += ' WHERE ' + this.idColumnEscaped(model) + ' = ' + id + ' AND ROWNUM <= 1';
-  } else {
-    sql += ' WHERE ' + this.idColumnEscaped(model) + ' IS NULL AND ROWNUM <= 1';
-  }
-
-  this.query(sql, function (err, data) {
-    if (err) return callback(err);
-    callback(null, data.length === 1);
-  });
-};
-
-/**
- * Find a model instance by id
- * @param {String} model The model name
- * @param {*} id The id value
- * @param {Function} [callback] The callback function
- */
-Oracle.prototype.find = function find(model, id, callback) {
-  var sql = 'SELECT * FROM ' +
-    this.tableEscaped(model);
-
-  if (id) {
-    var idVal = this.toDatabase(this._models[model].properties[this.idName(model)], id);
-    sql += ' WHERE ' + this.idColumnEscaped(model) + ' = ' + idVal + ' AND ROWNUM <= 1';
-  }
-  else {
-    sql += ' WHERE ' + this.idColumnEscaped(model) + ' IS NULL AND ROWNUM <= 1';
-  }
-
-  this.query(sql, function (err, data) {
-    if (!data || data.length === 0) {
-      data = [null];
-    }
-    callback(err, this.fromDatabase(model, data[0]));
-  }.bind(this));
-};
-
-/*!
- * Discover the properties from a table
- * @param {String} model The model name
- * @param {Function} cb The callback function
- */
-function getTableStatus(model, cb) {
-  function decoratedCallback(err, data) {
-    if (err) {
-      console.error(err);
-    }
-    if (!err) {
-      data.forEach(function (field) {
-        field.type = mapOracleDatatypes(field.type);
-      });
-    }
-    cb(err, data);
-  }
-
-  this.query('SELECT column_name AS "column", data_type AS "type", nullable AS "nullable"' // , data_default AS "Default"'
-    + ' FROM "SYS"."USER_TAB_COLUMNS" WHERE table_name=\'' + this.table(model) + '\'', decoratedCallback);
-
-}
-
-/*!
- * Create a SQL statement that supports pagination
- * @param {String} sql The SELECT statement that supports pagination
- * @param {String} orderBy The 'order by' columns
- * @param {Object} options options
- * @returns {String} The SQL statement
- */
-function paginateSQL(sql, orderBy, options) {
-  var pagination = getPagination(options);
-  orderBy = orderBy || '1';
-  if (pagination.length) {
-    return 'SELECT * FROM (SELECT ROW_NUMBER() OVER (ORDER BY ' + orderBy + ') R, '
-      + sql.substring(7) + ') WHERE ' + pagination.join(' AND ');
-  } else {
-    return sql;
-  }
-}
-
-/*!
- * Build sql for listing tables
- * @param {Object} options {all: for all owners, owner: for a given owner}
- * @returns {String} The sql statement
- */
-function queryTables(options) {
-  var sqlTables = null;
-  var owner = options.owner || options.schema;
-
-  if (options.all && !owner) {
-    sqlTables = paginateSQL('SELECT \'table\' AS "type", table_name AS "name", owner AS "owner"'
-      + ' FROM all_tables', 'owner, table_name', options);
-  } else if (owner) {
-    sqlTables = paginateSQL('SELECT \'table\' AS "type", table_name AS "name", owner AS "owner"'
-      + ' FROM all_tables WHERE owner=\'' + owner + '\'', 'owner, table_name', options);
-  } else {
-    sqlTables = paginateSQL('SELECT \'table\' AS "type", table_name AS "name",'
-      + ' SYS_CONTEXT(\'USERENV\', \'SESSION_USER\') AS "owner" FROM user_tables',
-      'table_name', options);
-  }
-  return sqlTables;
-}
-
-/*!
- * Build sql for listing views
- * @param {Object} options {all: for all owners, owner: for a given owner}
- * @returns {String} The sql statement
- */
-function queryViews(options) {
-  var sqlViews = null;
-  if (options.views) {
-
-    var owner = options.owner || options.schema;
-
-    if (options.all && !owner) {
-      sqlViews = paginateSQL('SELECT \'view\' AS "type", view_name AS "name",'
-        + ' owner AS "owner" FROM all_views',
-        'owner, view_name', options);
-    } else if (owner) {
-      sqlViews = paginateSQL('SELECT \'view\' AS "type", view_name AS "name",'
-        + ' owner AS "owner" FROM all_views WHERE owner=\'' + owner + '\'',
-        'owner, view_name', options);
-    } else {
-      sqlViews = paginateSQL('SELECT \'view\' AS "type", view_name AS "name",'
-        + ' SYS_CONTEXT(\'USERENV\', \'SESSION_USER\') AS "owner" FROM user_views',
-        'view_name', options);
-    }
-  }
-  return sqlViews;
-}
-
-/**
- * Discover model definitions.
- * Example of callback function return value:
- * ```js
- * {type: 'table', name: 'INVENTORY', owner: 'STRONGLOOP' }
- * {type: 'table', name: 'LOCATION', owner: 'STRONGLOOP' }
- * {type: 'view', name: 'INVENTORY_VIEW', owner: 'STRONGLOOP' }
- *```
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @param {Function} [cb] The callback function.
- * @private
- */
-Oracle.prototype.discoverModelDefinitions = function (options, cb) {
-  if (!cb && typeof options === 'function') {
-    cb = options;
-    options = {};
-  }
-  options = options || {};
-
-  var self = this;
-  var calls = [function (callback) {
-    self.query(queryTables(options), callback);
-  }];
-
-  if (options.views) {
-    calls.push(function (callback) {
-      self.query(queryViews(options), callback);
-    });
-  }
-  async.parallelLimit(calls, this.parallelLimit, function (err, data) {
-    if (err) {
-      cb(err, data);
-    } else {
-      var merged = [];
-      merged = merged.concat(data.shift());
-      if (data.length) {
-        merged = merged.concat(data.shift());
-      }
-      cb(err, merged);
-    }
-  });
-};
-
-/**
- * Discover the model definitions synchronously.
- * 
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @private
- */
-Oracle.prototype.discoverModelDefinitionsSync = function (options) {
-  options = options || {};
-  var sqlTables = queryTables(options);
-  var tables = this.querySync(sqlTables);
-  var sqlViews = queryViews(options);
-  if (sqlViews) {
-    var views = this.querySync(sqlViews);
-    tables = tables.concat(views);
-  }
-  return tables;
-};
-
-/*!
- * Normalize the arguments
- * @param {String} table The table name
- * @param {Object} [options] The options object
- * @param {Function} [cb] The callback function
- */
-function getArgs(table, options, cb) {
-  if ('string' !== typeof table || !table) {
-    throw new Error('table is a required string argument: ' + table);
-  }
-  options = options || {};
-  if (!cb && 'function' === typeof options) {
-    cb = options;
-    options = {};
-  }
-  if (typeof options !== 'object') {
-    throw new Error('options must be an object: ' + options);
-  }
-  return {
-    owner: options.owner || options.schema,
-    table: table,
-    options: options,
-    cb: cb
-  };
-}
-
-/*!
- * Build the sql statement to query columns for a given table
- * @param {String} owner The DB owner/schema name
- * @param {String} table The table name
- * @returns {String} The sql statement
- */
-function queryColumns(owner, table) {
-  var sql = null;
-  if (owner) {
-    sql = paginateSQL('SELECT owner AS "owner", table_name AS "tableName",'
-      + ' column_name AS "columnName", data_type AS "dataType",'
-      + ' data_length AS "dataLength", data_precision AS "dataPrecision",'
-      + ' data_scale AS "dataScale", nullable AS "nullable"'
-      + ' FROM all_tab_columns'
-      + ' WHERE owner=\'' + owner + '\''
-      + (table ? ' AND table_name=\'' + table + '\'' : ''),
-      'table_name, column_id', {});
-  } else {
-    sql = paginateSQL('SELECT SYS_CONTEXT(\'USERENV\', \'SESSION_USER\') AS "owner",'
-      + ' table_name AS "tableName", column_name AS "columnName", data_type AS "dataType",'
-      + ' data_length AS "dataLength", data_precision AS "dataPrecision",'
-      + ' data_scale AS "dataScale", nullable AS "nullable"'
-      + ' FROM user_tab_columns'
-      + (table ? ' WHERE table_name=\'' + table + '\'' : ''),
-      'table_name, column_id', {});
-  }
-  return sql;
-}
-
-/**
- * Discover model properties from a table.  Returns an array of columns for the specified table.
- *
- * @param {String} table The table name
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @param {Function} [cb] The callback function
- * 
- * ```js
- * { owner: 'STRONGLOOP',
- *   tableName: 'PRODUCT',
- *   columnName: 'ID',
- *   dataType: 'VARCHAR2',
- *   dataLength: 20,
- *   nullable: 'N',
- *   type: 'String'
- * }
- * { owner: 'STRONGLOOP',
- *   tableName: 'PRODUCT',
- *   columnName: 'NAME',
- *   dataType: 'VARCHAR2',
- *   dataLength: 64,
- *   nullable: 'Y',
- *   type: 'String'
- * }
- *```
- * @private
- */
-Oracle.prototype.discoverModelProperties = function (table, options, cb) {
-  var args = getArgs(table, options, cb);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-  cb = args.cb;
-
-  var sql = queryColumns(owner, table);
-  var callback = function (err, results) {
-    if (err) {
-      cb(err, results);
-    } else {
-      results.map(function (r) {
-        r.type = oracleDataTypeToJSONType(r.dataType, r.dataLength);
-      });
-      cb(err, results);
-    }
-  };
-  this.query(sql, callback);
-}
-
-/**
- * Discover model properties from a table synchronously.  See example return value for discoverModelProperties().
- * 
- * @param {String} table The table name
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @private
- * 
- */
-Oracle.prototype.discoverModelPropertiesSync = function (table, options) {
-  var args = getArgs(table, options);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-
-  var sql = queryColumns(owner, table);
-  var results = this.querySync(sql);
-  results.map(function (r) {
-    r.type = oracleDataTypeToJSONType(r.dataType, r.dataLength);
-  });
-  return results;
-}
-
-/*!
- * Build the sql statement for querying primary keys of a given table
- * @param owner
- * @param table
- * @returns {String}
- */
-// http://docs.oracle.com/javase/6/docs/api/java/sql/DatabaseMetaData.html#
-// getPrimaryKeys(java.lang.String, java.lang.String, java.lang.String)
-function queryForPrimaryKeys(owner, table) {
-  var sql = 'SELECT uc.owner AS "owner", '
-    + 'uc.table_name AS "tableName", col.column_name AS "columnName",'
-    + ' col.position AS "keySeq", uc.constraint_name AS "pkName" FROM'
-    + (owner ? ' ALL_CONSTRAINTS uc, ALL_CONS_COLUMNS col' : ' USER_CONSTRAINTS uc, USER_CONS_COLUMNS col')
-    + ' WHERE uc.constraint_type=\'P\' AND uc.constraint_name=col.constraint_name';
-
-  if (owner) {
-    sql += ' AND uc.owner=\'' + owner + '\'';
-  }
-  if (table) {
-    sql += ' AND uc.table_name=\'' + table + '\'';
-  }
-  sql += ' ORDER BY uc.owner, col.constraint_name, uc.table_name, col.position';
-  return sql;
-}
-
-/**
- * Discover primary keys for specified table. Returns an array of primary keys for the specified table.
- * Example return value:
- * ```js
- *         { owner: 'STRONGLOOP',
- *           tableName: 'INVENTORY',
- *           columnName: 'PRODUCT_ID',
- *           keySeq: 1,
- *           pkName: 'ID_PK' }
- *         { owner: 'STRONGLOOP',
- *           tableName: 'INVENTORY',
- *           columnName: 'LOCATION_ID',
- *           keySeq: 2,
- *          pkName: 'ID_PK' }
- *``` 
- * 
- * @param {String} table The table name
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @param {Function} [cb] The callback function
- * @private
- */
-Oracle.prototype.discoverPrimaryKeys = function (table, options, cb) {
-  var args = getArgs(table, options, cb);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-  cb = args.cb;
-
-  var sql = queryForPrimaryKeys(owner, table);
-  this.query(sql, cb);
-}
-
-/**
- * Discover primary keys synchronously for specified table.  See example return value for discoverPrimaryKeys().
- * 
- * @param {String} table
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @private
- */
-Oracle.prototype.discoverPrimaryKeysSync = function (table, options) {
-  var args = getArgs(table, options);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-
-  var sql = queryForPrimaryKeys(owner, table);
-  return this.querySync(sql);
-}
-
-/*!
- * Build the sql statement for querying foreign keys of a given table
- * @param {String} owner The DB owner/schema name
- * @param {String} table The table name
- * @returns {String} The SQL statement to find foreign keys
- */
-function queryForeignKeys(owner, table) {
-  var sql =
-    'SELECT a.owner AS "fkOwner", a.constraint_name AS "fkName", a.table_name AS "fkTableName",'
-      + ' a.column_name AS "fkColumnName", a.position AS "keySeq",'
-      + ' jcol.owner AS "pkOwner", jcol.constraint_name AS "pkName",'
-      + ' jcol.table_name AS "pkTableName", jcol.column_name AS "pkColumnName"'
-      + ' FROM'
-      + ' (SELECT'
-      + ' uc.owner, uc.table_name, uc.constraint_name, uc.r_constraint_name, col.column_name, col.position'
-      + ' FROM'
-      + (owner ? ' ALL_CONSTRAINTS uc, ALL_CONS_COLUMNS col' : ' USER_CONSTRAINTS uc, USER_CONS_COLUMNS col')
-      + ' WHERE'
-      + ' uc.constraint_type=\'R\' and uc.constraint_name=col.constraint_name';
-  if (owner) {
-    sql += ' AND uc.owner=\'' + owner + '\'';
-  }
-  if (table) {
-    sql += ' AND uc.table_name=\'' + table + '\'';
-  }
-  sql += ' ) a'
-    + ' INNER JOIN'
-    + ' USER_CONS_COLUMNS jcol'
-    + ' ON'
-    + ' a.r_constraint_name=jcol.constraint_name';
-  return sql;
-}
-
-/**
- * Discover foreign keys for specified table.  Example return value:
- * ```js
- * { fkOwner: 'STRONGLOOP',
- *   fkName: 'PRODUCT_FK',
- *   fkTableName: 'INVENTORY',
- *   fkColumnName: 'PRODUCT_ID',
- *   keySeq: 1,
- *   pkOwner: 'STRONGLOOP',
- *   pkName: 'PRODUCT_PK',
- *   pkTableName: 'PRODUCT',
- *   pkColumnName: 'ID' }
- * ```
- * 
- * @param {String} table The table name
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @param {Function} [cb] The callback function
- * @private
- */
-Oracle.prototype.discoverForeignKeys = function (table, options, cb) {
-  var args = getArgs(table, options, cb);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-  cb = args.cb;
-
-  var sql = queryForeignKeys(owner, table);
-  this.query(sql, cb);
-};
-
-/**
- * Discover foreign keys synchronously for a given table.  See example return value for discoverForeignKeys().
- * 
- * @param {String} table The table name
- * @options {Object} options Options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @prop {Number} limit Maximimum number of results to return.  NOTE: This was not in .md doc, but is included in tests.
- * @private
- */
-Oracle.prototype.discoverForeignKeysSync = function (table, options) {
-  var args = getArgs(table, options);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-
-  var sql = queryForeignKeys(owner, table);
-  return this.querySync(sql);
-};
-
-/*!
- * Retrieves a description of the foreign key columns that reference the given
- * table's primary key columns (the foreign keys exported by a table).
- * They are ordered by fkTableOwner, fkTableName, and keySeq.
- * @param {String} owner The DB owner/schema name
- * @param {String} table The table name
- * @returns {String} The SQL statement
- */
-function queryExportedForeignKeys(owner, table) {
-  var sql = 'SELECT a.constraint_name AS "fkName", a.owner AS "fkOwner", a.table_name AS "fkTableName",'
-    + ' a.column_name AS "fkColumnName", a.position AS "keySeq",'
-    + ' jcol.constraint_name AS "pkName", jcol.owner AS "pkOwner",'
-    + ' jcol.table_name AS "pkTableName", jcol.column_name AS "pkColumnName"'
-    + ' FROM'
-    + ' (SELECT'
-    + ' uc1.table_name, uc1.constraint_name, uc1.r_constraint_name, col.column_name, col.position, col.owner'
-    + ' FROM'
-    + (owner ? ' ALL_CONSTRAINTS uc, ALL_CONSTRAINTS uc1, ALL_CONS_COLUMNS col'
-    : ' USER_CONSTRAINTS uc, USER_CONSTRAINTS uc1, USER_CONS_COLUMNS col')
-    + ' WHERE'
-    + ' uc.constraint_type=\'P\' and uc1.r_constraint_name = uc.constraint_name and uc1.constraint_type = \'R\''
-    + ' and uc1.constraint_name=col.constraint_name';
-  if (owner) {
-    sql += ' and col.owner=\'' + owner + '\'';
-  }
-  if (table) {
-    sql += ' and uc.table_Name=\'' + table + '\'';
-  }
-  sql += ' ) a'
-    + ' INNER JOIN'
-    + ' USER_CONS_COLUMNS jcol'
-    + ' ON'
-    + ' a.r_constraint_name=jcol.constraint_name'
-    + ' order by a.owner, a.table_name, a.position';
-
-  return sql;
-}
-
-/**
- * Discover foreign keys that reference to the primary key of this table.
- * Example return value:
- * ```js
- * { fkName: 'PRODUCT_FK',
- *   fkOwner: 'STRONGLOOP',
- *   fkTableName: 'INVENTORY',
- *   fkColumnName: 'PRODUCT_ID',
- *   keySeq: 1,
- *   pkName: 'PRODUCT_PK',
- *   pkOwner: 'STRONGLOOP',
- *   pkTableName: 'PRODUCT',
- *   pkColumnName: 'ID' }
- *   ````
- * @param {String} table The table name
- * @options {Object} options The options for discovery
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @param {Function} [cb] The callback function
- * @private
- */
-Oracle.prototype.discoverExportedForeignKeys = function (table, options, cb) {
-  var args = getArgs(table, options, cb);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-  cb = args.cb;
-
-  var sql = queryExportedForeignKeys(owner, table);
-  this.query(sql, cb);
-};
-
-/**
- * Discover foreign keys synchronously for a given table; see example return value for discoverExportedForeignKeys().
- * @param {String} owner The DB owner/schema name
- * @options {Object} options The options for discovery; see below.
- * @prop {Boolean} all If true, include tables/views from all schemas/owners.
- * @prop {String} owner/schema The schema/owner name.  QUESTION:  What is the actual name of this property: 'owner' or 'schema'?  Or can it be either?
- * @prop {Boolean} views If true, include views.
- * @returns {*}
- * @private
- */
-Oracle.prototype.discoverExportedForeignKeysSync = function (table, options) {
-  var args = getArgs(table, options);
-  var owner = args.owner;
-  table = args.table;
-  options = args.options;
-
-  var sql = queryExportedForeignKeys(owner, table);
-  return this.querySync(sql);
-};
-
-/**
- * Perform autoupdate for the given models
- * @param {String[]} [models] A model name or an array of model names. If not
- * present, apply to all models
- * @param {Function} [cb] The callback function
- */
-Oracle.prototype.autoupdate = function(models, cb) {
-  var self = this;
-  if ((!cb) && ('function' === typeof models)) {
-    cb = models;
-    models = undefined;
-  }
-  // First argument is a model name
-  if ('string' === typeof models) {
-    models = [models];
-  }
-
-  models = models || Object.keys(this._models);
-
-  async.eachLimit(models, this.parallelLimit, function(model, done) {
-    if (!(model in self._models)) {
-      return process.nextTick(function() {
-        done(new Error('Model not found: ' + model));
-      });
-    }
-    getTableStatus.call(self, model, function(err, fields) {
-      if (!err && fields.length) {
-        self.alterTable(model, fields, done);
-      } else {
-        self.createTable(model, done);
-      }
-    });
-  }, cb);
-};
-
-/*!
- * Check if the models exist
- * @param {String[]} [models] A model name or an array of model names. If not
- * present, apply to all models
- * @param {Function} [cb] The callback function
- */
-Oracle.prototype.isActual = function(models, cb) {
-  var self = this;
-
-  if ((!cb) && ('function' === typeof models)) {
-    cb = models;
-    models = undefined;
-  }
-  // First argument is a model name
-  if ('string' === typeof models) {
-    models = [models];
-  }
-
-  models = models || Object.keys(this._models);
-
-  var changes = [];
-  async.eachLimit(models, this.parallelLimit, function(model, done) {
-    getTableStatus.call(self, model, function(err, fields) {
-      changes = changes.concat(getAddModifyColumns.call(self, model, fields));
-      changes = changes.concat(getDropColumns.call(self, model, fields));
-      done(err);
-    });
-  }, function done(err) {
-    if (err) {
-      return cb && cb(err);
-    }
-    var actual = (changes.length === 0);
-    cb && cb(null, actual);
-  });
-};
-
-/**
- * Alter the table for the given model
- * @param {String} model The model name
- * @param {Object[]} actualFields Actual columns in the table
- * @param {Function} [cb] The callback function
- */
-Oracle.prototype.alterTable = function (model, actualFields, cb) {
-  var self = this;
-  var pendingChanges = getAddModifyColumns.call(self, model, actualFields);
-  if (pendingChanges.length > 0) {
-    applySqlChanges.call(self, model, pendingChanges, function (err, results) {
-      var dropColumns = getDropColumns.call(self, model, actualFields);
-      if (dropColumns.length > 0) {
-        applySqlChanges.call(self, model, dropColumns, cb);
-      } else {
-        cb && cb(err, results);
-      }
-    });
-  } else {
-    var dropColumns = getDropColumns.call(self, model, actualFields);
-    if (dropColumns.length > 0) {
-      applySqlChanges.call(self, model, dropColumns, cb);
-    } else {
-      cb && process.nextTick(cb.bind(null, null, []));
-    }
-  }
-};
-
-function getAddModifyColumns(model, actualFields) {
-  var sql = [];
-  var self = this;
-  sql = sql.concat(getColumnsToAdd.call(self, model, actualFields));
-  sql = sql.concat(getPropertiesToModify.call(self, model, actualFields));
-  return sql;
-}
-
-function getDropColumns(model, actualFields) {
-  var sql = [];
-  var self = this;
-  sql = sql.concat(getColumnsToDrop.call(self, model, actualFields));
-  return sql;
-}
-
-function getColumnsToAdd(model, actualFields) {
-  var self = this;
-  var m = self._models[model];
-  var propNames = Object.keys(m.properties);
-  var sql = [];
-  propNames.forEach(function (propName) {
-    if (self.id(model, propName)) {
-      return;
-    }
-    var found = searchForPropertyInActual.call(self, model, self.column(model, propName), actualFields);
-    if (!found && propertyHasNotBeenDeleted.call(self, model, propName)) {
-      sql.push(addPropertyToActual.call(self, model, propName));
-    }
-  });
-  if (sql.length > 0) {
-    sql = ['ADD', '(' + sql.join(',') + ')'];
-  }
-  return sql;
-}
-
-function addPropertyToActual(model, propName) {
-  var self = this;
-  var sqlCommand = self.columnEscaped(model, propName)
-    + ' ' + self.columnDataType(model, propName)
-    + (propertyCanBeNull.call(self, model, propName) ? "" : " NOT NULL");
-  return sqlCommand;
-}
-
-function searchForPropertyInActual(model, propName, actualFields) {
-  var self = this;
-  var found = false;
-  actualFields.forEach(function (f) {
-    if (f.column === self.column(model, propName)) {
-      found = f;
-      return;
-    }
-  });
-  return found;
-}
-
-function getPropertiesToModify(model, actualFields) {
-  var self = this;
-  var sql = [];
-  var m = self._models[model];
-  var propNames = Object.keys(m.properties);
-  var found;
-  propNames.forEach(function(propName) {
-    if (self.id(model, propName)) {
-      return;
-    }
-    found = searchForPropertyInActual.call(self, model, propName, actualFields);
-    if (found && propertyHasNotBeenDeleted.call(self, model, propName)) {
-      var column = self.columnEscaped(model, propName);
-      var clause = '';
-      if (datatypeChanged(propName, found)) {
-        clause = column + ' ' +
-          modifyDatatypeInActual.call(self, model, propName);
-      }
-      if (nullabilityChanged(propName, found)) {
-        if (!clause) {
-          clause = column;
-        }
-        clause = clause + ' ' +
-          modifyNullabilityInActual.call(self, model, propName);
-      }
-      if (clause) {
-        sql.push(clause);
+      if (filter.order) {
+        selectStmt.merge(this.buildOrderBy(model, filter.order));
       }
     }
-  });
 
-  if (sql.length > 0) {
-    sql = ['MODIFY', '(' + sql.join(',') + ')'];
   }
-  return sql;
-
-  function datatypeChanged(propName, oldSettings) {
-    var newSettings = m.properties[propName];
-    if (!newSettings) {
-      return false;
-    }
-    return oldSettings.type.toUpperCase() !== self.columnDataType(model, propName);
-  }
-
-  function nullabilityChanged(propName, oldSettings) {
-    var newSettings = m.properties[propName];
-    if (!newSettings) {
-      return false;
-    }
-    var changed = false;
-    if (oldSettings.nullable === 'Y' && !isNullable(newSettings)) {
-      changed = true;
-    }
-    if (oldSettings.nullable === 'N' && isNullable(newSettings)) {
-      changed = true;
-    }
-    return changed;
-  }
-}
-
-function modifyDatatypeInActual(model, propName) {
-  var self = this;
-  var sqlCommand = self.columnDataType(model, propName);
-  return sqlCommand;
-}
-
-function modifyNullabilityInActual(model, propName) {
-  var self = this;
-  var sqlCommand = '';
-  if (propertyCanBeNull.call(self, model, propName)) {
-    sqlCommand = sqlCommand + "NULL";
-  } else {
-    sqlCommand = sqlCommand + "NOT NULL";
-  }
-  return sqlCommand;
-}
-
-function getColumnsToDrop(model, actualFields) {
-  var self = this;
-  var sql = [];
-  actualFields.forEach(function (actualField) {
-    if (self.idColumn(model) === actualField.column) {
-      return;
-    }
-    if (actualFieldNotPresentInModel(actualField, model)) {
-      sql.push(self.escapeName(actualField.column));
-    }
-  });
-  if (sql.length > 0) {
-    sql = ['DROP', '(' + sql.join(',') + ')'];
-  }
-  return sql;
-
-  function actualFieldNotPresentInModel(actualField, model) {
-    return !(self.propertyName(model, actualField.column));
-  }
-}
-
-function applySqlChanges(model, pendingChanges, cb) {
-  var self = this;
-  if (pendingChanges.length) {
-    var thisQuery = 'ALTER TABLE ' + self.tableEscaped(model);
-    var ranOnce = false;
-    pendingChanges.forEach(function (change) {
-      if (ranOnce) {
-        thisQuery = thisQuery + ' ';
-      }
-      thisQuery = thisQuery + ' ' + change;
-      ranOnce = true;
-    });
-    self.query(thisQuery, cb);
-  }
-}
-
-/*!
- * Build a list of columns for the given model
- * @param {String} model The model name
- * @returns {String}
- */
-Oracle.prototype.propertiesSQL = function (model) {
-  var self = this;
-  var sql = [];
-  var pks = this.idNames(model).map(function (i) {
-    return self.columnEscaped(model, i);
-  });
-  Object.keys(this._models[model].properties).forEach(function (prop) {
-    var colName = self.columnEscaped(model, prop);
-    sql.push(colName + ' ' + self.propertySettingsSQL(model, prop));
-  });
-  if (pks.length > 0) {
-    sql.push('PRIMARY KEY(' + pks.join(',') + ')');
-  }
-  return sql.join(',\n  ');
-
-};
-
-/*!
- * Build settings for the model property
- * @param {String} model The model name
- * @param {String} propName The property name
- * @returns {*|string}
- */
-Oracle.prototype.propertySettingsSQL = function (model, propName) {
-  var self = this;
-  var result = self.columnDataType(model, propName);
-  if (!propertyCanBeNull.call(self, model, propName)) result = result + ' NOT NULL';
-  return result;
-};
-
-Oracle.prototype._isIdGenerated = function (model) {
-  var idNames = this.idNames(model);
-  if (!idNames) {
-    return false;
-  }
-  var idName = idNames[0];
-  var id = this._models[model].properties[idName];
-  var idGenerated = idNames.length > 1 ? false : id && id.generated;
-  return idGenerated;
-};
-
-/**
- * Drop a table for the given model
- * @param {String} model The model name
- * @param {Function} [cb] The callback function
- */
-Oracle.prototype.dropTable = function (model, cb) {
-  var self = this;
-  var name = self.tableEscaped(model);
-  var seqName = self.escapeName(model + '_ID_SEQUENCE');
-
-  var count = 0;
-  var dropTableFun = function (callback) {
-    self.query('DROP TABLE ' + name, function (err, data) {
-      if (err && err.toString().indexOf('ORA-00054') >= 0) {
-        count++;
-        if (count <= 5) {
-          self.debug('Retrying ' + count + ': ' + err);
-          // Resource busy, try again
-          setTimeout(dropTableFun, 200 * Math.pow(count, 2));
-          return;
-        }
-      }
-      if (err && err.toString().indexOf('ORA-00942') >= 0) {
-        err = null; // Ignore it
-      }
-      callback(err, data);
-    });
-  };
-
-  var tasks = [dropTableFun];
-  if (this._isIdGenerated(model)) {
-    tasks.push(
-      function (callback) {
-        self.query('DROP SEQUENCE ' + seqName, function(err) {
-          if (err && err.toString().indexOf('ORA-02289') >= 0) {
-            err = null; // Ignore it
-          }
-          callback(err);
-        });
-      });
-    // Triggers will be dropped as part the drop table
-  }
-  async.series(tasks, cb);
-};
-
-/**
- * Create a table for the given model
- * @param {String} model The model name
- * @param {Function} [cb] The callback function
- */
-Oracle.prototype.createTable = function (model, cb) {
-  var self = this;
-  var name = self.tableEscaped(model);
-  var seqName = self.escapeName(model + '_ID_SEQUENCE');
-  var triggerName = self.escapeName(model + '_ID_TRIGGER');
-  var idName = self.idColumnEscaped(model);
-
-  var tasks = [
-    function (callback) {
-      self.query('CREATE TABLE ' + name + ' (\n  ' + self.propertiesSQL(model) + '\n)', callback);
-    }];
-
-  if (this._isIdGenerated(model)) {
-    tasks.push(
-      function (callback) {
-        self.query('CREATE SEQUENCE ' + seqName +
-          ' START WITH 1 INCREMENT BY 1 CACHE 100', callback);
-      });
-
-    tasks.push(
-      function(callback) {
-        self.query('CREATE OR REPLACE TRIGGER ' + triggerName +
-          ' BEFORE INSERT ON ' + name + ' FOR EACH ROW\n' +
-          'WHEN (new.' + idName + ' IS NULL)\n' +
-          'BEGIN\n' +
-          '  SELECT ' + seqName + '.NEXTVAL INTO :new.' +
-          idName + ' FROM dual;\n' +
-          'END;', callback);
-      });
-  }
-
-  async.series(tasks, cb);
+  return this.parameterize(selectStmt);
 };
 
 /**
@@ -1944,166 +508,9 @@ Oracle.prototype.disconnect = function disconnect(cb) {
   }
 };
 
-function isNullable(p) {
-  return !(p.required || p.id || p.nullable === false ||
-    p.allowNull === false || p['null'] === false);
-}
-
-function propertyCanBeNull(model, propName) {
-  var p = this._models[model].properties[propName];
-  return isNullable(p);
-}
-
-function escape(val) {
-  if (val === undefined || val === null) {
-    return 'NULL';
-  }
-
-  switch (typeof val) {
-    case 'boolean':
-      return (val) ? "'Y'" : "'N'";
-    case 'number':
-      return val + '';
-  }
-
-  if (typeof val === 'object') {
-    val = (typeof val.toISOString === 'function')
-      ? val.toISOString()
-      : val.toString();
-  }
-
-  val = val.replace(/[\0\n\r\b\t\\\'\"\x1a]/g, function (s) {
-    switch (s) {
-      case "\0":
-        return "\\0";
-      case "\n":
-        return "\\n";
-      case "\r":
-        return "\\r";
-      case "\b":
-        return "\\b";
-      case "\t":
-        return "\\t";
-      case "\x1a":
-        return "\\Z";
-      case "\'":
-        return "''"; // For oracle
-      case "\"":
-        return s; // For oracle
-      default:
-        return "\\" + s;
-    }
-  });
-  // return "q'#"+val+"#'";
-  return "'" + val + "'";
-}
-
-/*!
- * Find the column type for a given model property
- *
- * @param {String} model The model name
- * @param {String} property The property name
- * @returns {String} The column type
- */
-Oracle.prototype.columnDataType = function (model, property) {
-  var columnMetadata = this.columnMetadata(model, property);
-  var colType = columnMetadata && columnMetadata.dataType;
-  if (colType) {
-    colType = colType.toUpperCase();
-  }
-  var prop = this._models[model].properties[property];
-  if (!prop) {
-    return null;
-  }
-  var colLength = columnMetadata && columnMetadata.dataLength || prop.length;
-  if (colType) {
-    return colType + (colLength ? '(' + colLength + ')' : '');
-  }
-
-  switch (prop.type.name) {
-    default:
-    case 'String':
-    case 'JSON':
-      return 'VARCHAR2' + (colLength ? '(' + colLength + ')' : '(1024)');
-    case 'Text':
-      return 'VARCHAR2' + (colLength ? '(' + colLength + ')' : '(1024)');
-    case 'Number':
-      return 'NUMBER';
-    case 'Date':
-      return 'DATE';
-    case 'Timestamp':
-      return 'TIMESTAMP(3)';
-    case 'Boolean':
-      return 'CHAR(1)'; // Oracle doesn't have built-in boolean
-  }
-};
-
 Oracle.prototype.ping = function (cb) {
-  this.query('select count(*) as result from user_tables', [], cb);
+  this.execute('select count(*) as result from user_tables', [], cb);
 };
 
-/*!
- * Map oracle data types to json types
- * @param {String} oracleType
- * @param {Number} dataLength
- * @returns {String}
- */
-function oracleDataTypeToJSONType(oracleType, dataLength) {
-  var type = oracleType.toUpperCase();
-  switch (type) {
-    case 'CHAR':
-      if (dataLength === 1) {
-        // Treat char(1) as boolean
-        return 'Boolean';
-      } else {
-        return 'String';
-      }
-      break;
-    case 'VARCHAR':
-    case 'VARCHAR2':
-    case 'LONG VARCHAR':
-    case 'NCHAR':
-    case 'NVARCHAR2':
-      return 'String';
-    case 'LONG':
-    case 'BLOB':
-    case 'CLOB':
-    case 'NCLOB':
-      return 'Binary';
-    case 'NUMBER':
-    case 'INTEGER':
-    case 'DECIMAL':
-    case 'DOUBLE':
-    case 'FLOAT':
-    case 'BIGINT':
-    case 'SMALLINT':
-    case 'REAL':
-    case 'NUMERIC':
-    case 'BINARY_FLOAT':
-    case 'BINARY_DOUBLE':
-    case 'UROWID':
-    case 'ROWID':
-      return 'Number';
-    case 'DATE':
-    case 'TIMESTAMP':
-      return 'Date';
-    default:
-      return 'String';
-  }
-}
-
-function mapOracleDatatypes(typeName) {
-  //TODO there are a lot of synonymous type names that should go here-- this is just what i've run into so far
-  switch (typeName) {
-    case 'int4':
-      return 'NUMBER';
-    case 'bool':
-      return 'CHAR(1)';
-    default:
-      return typeName;
-  }
-}
-
-function propertyHasNotBeenDeleted(model, propName) {
-  return !!this._models[model].properties[propName];
-}
+require('./migration')(Oracle);
+require('./discovery')(Oracle);
